containers_build/auth-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
requests==2.31.0
mysql-connector-python==8.2.0
python-jose==3.3.0
gunicorn==21.2.0
cryptography==41.0.7
